Misc cleanup

+++ src/ray.h
@@ -35,7 +35,7 @@ public:
     inline const glm::vec3& dir() const { return mDir; }
     inline const glm::vec3& inverseDir() const { return mInverseDir; }
     inline glm::vec3 point(const float t) const { return mOrigin + mDir * t; }
-    inline short depth() const { return mDepth; }
+    inline unsigned depth() const { return mDepth; }
     inline bool shouldHitBackFaces() const { return mHitBack; }
     inline float ior() const { return mIor; }
     inline TYPE type() const { return mType; }
@@ -64,7 +64,7 @@ private:
     const TYPE mType;
     glm::vec3 mOrigin;
     glm::vec3 mDir, mInverseDir;
-    int mDepth;
+    unsigned mDepth;
     float mIor;
     float mMinT, mMaxT;
     const IPrimitive *mHitPrim;
